Add Ampersand Ligature Variant for `ss03`
The feature `ss03` changes the ampersand to the traditional style.
When this feature is enabled the restyling of the ligature form
`ampersand_ampersand.liga` is not used because it does not exist.

This commit adds the part glyph `_part.ampersand.ss03`, adds the
ligature glyph `ampersand_ampersand.liga.ss03`, and modifies the
`ss03` feature to use the newly added ligature when the feature is
enabled.

The part glyph was only created to be a component for the newly
added ligature, like `_part.ampersand` is for
`ampersand_ampersand.liga`.